Moves query_stats number of operations tests
Just moves the query_stats changes from `test_s3.py` to a new
`test_num_storage_operations.py`

And moves common `storage_bucket` fixture to conftest.
